Metagenerator/requirements.txt
Flask==2.3.2
pdfminer.six==20221105
python-docx==0.8.11
pytesseract==0.3.10
Pillow==9.5.0
requests==2.31.0
python-dotenv==1.0.0
